Update mock metadata result
