+++ packages/twenty-server/src/engine/workspace-manager/workspace-sync-metadata/constants/standard-field-ids.ts
@@ -453,11 +453,6 @@ export const WEBHOOK_STANDARD_FIELD_IDS = {
   secret: '20202020-97ce-410f-bff9-e9ccb038fb67',
 };
 
-export const WORKFLOW_EVENT_LISTENER_STANDARD_FIELD_IDS = {
-  eventName: '20202020-7318-4cf8-a6ac-2de75e3fd97d',
-  workflow: '20202020-4082-4641-8569-dc08d5365002',
-};
-
 export const WORKFLOW_AUTOMATED_TRIGGER_STANDARD_FIELD_IDS = {
   type: '20202020-3319-4234-a34c-3f92c1ab56e7',
   settings: '20202020-3319-4234-a34c-bac8f903de12',
@@ -585,7 +580,6 @@ export const STANDARD_OBJECT_FIELD_IDS = {
   view: VIEW_STANDARD_FIELD_IDS,
   webhook: WEBHOOK_STANDARD_FIELD_IDS,
   workflow: WORKFLOW_STANDARD_FIELD_IDS,
-  workflowEventListener: WORKFLOW_EVENT_LISTENER_STANDARD_FIELD_IDS,
   workflowRun: WORKFLOW_RUN_STANDARD_FIELD_IDS,
   workflowVersion: WORKFLOW_VERSION_STANDARD_FIELD_IDS,
   workspaceMember: WORKSPACE_MEMBER_STANDARD_FIELD_IDS,
